+++ .gitignore
@@ -19,5 +19,5 @@
 
 # Go workspace file
 go.work
-/gh-hubbub.exe
-/gh-hubbub
+/gh-reponark.exe
+/gh-reponark
