+++ filters/BoolModel.go
@@ -1,7 +1,7 @@
 package filters
 
 import (
-	"gh-hubbub/shared"
+	"gh-reponark/shared"
 
 	"github.com/charmbracelet/bubbles/v2/textinput"
 	tea "github.com/charmbracelet/bubbletea/v2"
