+++ users/user.go
@@ -1,4 +1,4 @@
-package user
+package users
 
 type User struct {
 	Name  string
